Fix branching logic in first java example
The branching logic on line 40 was backwards using weaviate 1.14.1 and the 3.2.0 java client library. This fixes things and the meta info on the running instance is reported properly.

Before this change I hit the code on line 45 which caused a null pointer exception because meta.getError() returns null in the non-error case.