Add favicon.ico
This is a tiny 1kb file that is a mini version of the Weaviate logo. It will appear whenever anyone accesses a Weaviate.io website from now on. 

Why are we doing it? nearly 10k people a month get a 404 when trying to access this file. Let's make that experience a bit better with a lovely Weaviate logo!